+++ src/lib/commonUtils.js
