+++ src/lib/components/EntranceWindow.svelte
